Created MealsList component
Refactored list of rendered meals into its
own Component file which will be used to
render meals on btoh the OverviewScreen
and FavoritesScreen.
