+++ react-native-coursework/section-6-meals-app/Meals/screens/MealsOverviewScreen.tsx
@@ -1,9 +1,9 @@
 import { useLayoutEffect } from "react";
-import { View, FlatList, StyleSheet, ListRenderItemInfo } from "react-native"
+import { View, FlatList } from "react-native"
 import type { NativeStackScreenProps } from '@react-navigation/native-stack';
 
 import { MEALS } from "../data/meal-data"
-import MealItem from "../components/MealItem";
+import MealsList from "../components/MealsList/MealsList";
 import Meal from "../models/meal";
 import { CATEGORIES } from "../data/meal-data";
 
@@ -33,33 +33,9 @@ function MealsOverviewScreen({ route, navigation }: Props) {
     
   }, [catId, navigation])
 
-  function renderMealItem(item: Meal) {
-    return <MealItem 
-            id={item.id}
-            title={item.title}
-            imageUrl={item.imageUrl}
-            duration={item.duration}
-            complexity={item.complexity}
-            affordability={item.affordability} 
-           />
-  }
+  
 
-  return (
-    <View style={styles.container}>
-      <FlatList 
-        data={displayedMeals}
-        keyExtractor={(item) => item.id}
-        renderItem={(itemData) => renderMealItem(itemData.item)}
-      />
-    </View>
-  )
+  return <MealsList items={displayedMeals}/>
 };
 
 export default MealsOverviewScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 16
-  }
-})
